Move VerfahrenTileSkeleton into /verfahren
