+++ app/components/verfahren/VerfahrenTileSkeleton.static.tsx
