+++ app/routes/verfahren.$id.bearbeiten.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 import { Await, Link, LoaderFunctionArgs, useLoaderData } from "react-router";
 import Alert from "~/components/Alert";
-import VerfahrenTileSkeleton from "~/components/VerfahrenTileSkeleton.static";
+import VerfahrenTileSkeleton from "~/components/verfahren/VerfahrenTileSkeleton.static";
 import fetchVerfahrenById from "~/domains/verfahren/fetchVerfahrenById.server";
 import { authContext, authMiddleware } from "~/middleware/auth.server";
 import { useTranslations } from "~/services/translations/context";
