+++ app/routes/verfahren.$id.tsx
@@ -1,8 +1,8 @@
 import { Suspense } from "react";
 import { Await, Link, LoaderFunctionArgs, useLoaderData } from "react-router";
 import Alert from "~/components/Alert";
-import VerfahrenTileSkeleton from "~/components/VerfahrenTileSkeleton.static";
 import VerfahrenTile from "~/components/verfahren/VerfahrenTile";
+import VerfahrenTileSkeleton from "~/components/verfahren/VerfahrenTileSkeleton.static";
 import fetchVerfahrenById from "~/domains/verfahren/fetchVerfahrenById.server";
 import { authContext, authMiddleware } from "~/middleware/auth.server";
 import { useTranslations } from "~/services/translations/context";
